tests/wpt/web-platform-tests/dom/collections/HTMLCollection-supported-property-names.html

<!doctype html>
<meta charset=utf-8>
<link rel=help href=https://dom.spec.whatwg.org/#interface-htmlcollection>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<div id=log></div>

<!-- with no attribute -->
<span></span>

<!-- with `id` attribute -->
<span id=''></span>
<span id='some-id'></span>
<span id='some-id'></span><!-- to ensure no duplicates -->

<!-- with `name` attribute -->
<span name=''></span>
<span name='some-name'></span>
<span name='some-name'></span><!-- to ensure no duplicates -->

<!-- with `name` and `id` attribute -->
<span id='another-id' name='another-name'></span>

<!-- TODO: test element with `name` that is not in HTML namespace -->

<script>
test(function () {
  var elements = document.getElementsByTagName("span");
  assert_array_equals(
    Object.getOwnPropertyNames(elements).sort(),
    ['0', '1', '2', '3', '4', '5', '6', '7', 'another-id', 'another-name', 'some-id', 'some-name']
  );
}, 'Object.getOwnPropertyNames on HTMLCollection');
</script>
